refactor(app): move an odd fixture
