+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupParameters/ChooseNumber.tsx
@@ -11,8 +11,8 @@ import {
   TYPOGRAPHY,
 } from '@opentrons/components'
 
-import { useToaster } from '../../../ToasterOven'
-import { ChildNavigation } from '../../../ChildNavigation'
+import { useToaster } from '/app/organisms/ToasterOven'
+import { ChildNavigation } from '/app/organisms/ChildNavigation'
 import { NumericalKeyboard } from '/app/atoms/SoftwareKeyboard'
 import type { NumberParameter } from '@opentrons/shared-data'
