+++ app/src/pages/ODD/ProtocolDetails/__tests__/ProtocolDetails.test.tsx
@@ -17,6 +17,7 @@ import { useHardwareStatusText } from '/app/organisms/ODD/RobotDashboard/hooks'
 import { useOffsetCandidatesForAnalysis } from '/app/organisms/ApplyHistoricOffsets/hooks/useOffsetCandidatesForAnalysis'
 import { useRunTimeParameters } from '/app/pages/Desktop/Protocols/hooks'
 import { ProtocolSetupParameters } from '/app/organisms/ODD/ProtocolSetup/ProtocolSetupParameters'
+import { mockRunTimeParameterData } from '/app/organisms/ODD/ProtocolSetup/__fixtures__'
 import { formatTimeWithUtcLabel } from '/app/resources/runs'
 import { useMissingProtocolHardware } from '/app/transformations/commands'
 import { ProtocolDetails } from '..'
@@ -24,7 +25,6 @@ import { Deck } from '../Deck'
 import { Hardware } from '../Hardware'
 import { Labware } from '../Labware'
 import { Parameters } from '../Parameters'
-import { mockRunTimeParameterData } from '../fixtures'
 
 import type { HostConfig } from '@opentrons/api-client'
